add anonymity set definitions
this commit adds 4 related definitions of anonymity sets, and the
supporting pages for their concepts

and under the assumption of sybil resistance, and depending on the kind
of threat model the user is concerned about:

- adversary has access to public information (inter-tx anonymity set
  through sub-transaction model & equivalence)
- the adversary is an intermediary (path like anonymity set)

these roughly correspond to Diaz et al's global passive external adversary and
global active external adversary, but it's only roughly, because
"global" in the passive case is not as powerful as in e.g. a
communication network system model, at least with respect to on chain
data since this data is broadcast/publicly observable.

the same data can be interpreted differently to estimate the user's risk
of deanonymization.

the path like anonymity set definition is a strict lower bound on the
inter-transaction definitions. a cost function that accounts for this
metric is attempting to address the most